3dom and iframes
Use case: A page contains a model-viewer instance but the functionality to control
that instance through 3dom scripts is emedded in an iframe.

* Modified ThreeDOMExecutionContext to create the worker using a factory method.

* 2 new classes that inherit from ThreeDOMExecutionContext and override the worker
creation method factory method.
  - HostThreeDOMExecutionContext: For the host page that includes the iframe. The
returned worker is fake and just used to intercept the passing of the MessagePort.
The port is transferred not to the worker in the host page but to a worker in the
iframe page. This class handles all the handshake necessary with the iframe.
  - IFrameThreeDOMExecutionContext: For the iframe page. The returned worker is
an actual worker but it waits to hand over the port until it is received from the
host page.

* The scene-graph feature exposes new method: setThreeDOMExecutionContext.

* New demo to show the iframe functionality.