Upgrade Filament and use its TypeScript bindings. (#1046)
* Minor update to fidelity tools README.

* Add gl-matrix dependency for Filament.

We do not actually use this library so it does not increase the size of
the build output. However it is a required dependency for Filament's
TypeScript bindings.

* Upgrade Filament and use its TypeScript bindings.

This commit updates API usage and required Filament version but does not
update golden images.

* Remove "import *" for Filament, etc.

Since Filament is bundled into an emscripten-generated CommonJS module,
it's tricky to import individual types. To achieve this, I had to use
the `external-globals` rollup plugin. This plugin allows us to honor the
import statements for type checking, but to ignore them at run time.

* Create alias for Filament::fetch to fix build.

* Upgrade Filament release candidate.

* filament-viewer: clean up imports, package deps.